BAH-4858 | Refactor. forlder structure
