+++ packages/bahmni-widgets/src/tasks/actions/index.ts
@@ -1,0 +1,2 @@
+export { default as TaskActions } from './TaskActions';
+export { handleTaskAction } from './actionHandlers';
