+++ packages/bahmni-widgets/src/tasks/views/ViewFormData/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ViewFormData';
