+++ packages/bahmni-widgets/src/tasks/views/index.ts
@@ -1,0 +1,3 @@
+export { default as TaskViewResults } from './TaskViewResults';
+export { handleTaskView } from './viewHandlers';
+export { default as ViewFormData } from './ViewFormData';
